Improved SignalSleeve Pipeline
Updated blue scalar bounds to work on the field lighting.
Added variables to hold colors.
Added a horizon and recreated the way contours are handled. First the contours list is sorted in descending order by width, then the list is looped through to find the largest contours which has a center point below the horizon. 

Wrap text function to help hide the weird ternary math.